backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
passlib[argon2]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==4.1.2
alembic==1.13.2
pycountry==24.6.1      
python-dotenv==1.0.1
